fix leaving on two devices

+++ resources/js/Pages/Game.jsx
@@ -5,16 +5,13 @@ import { AiOutlineSend } from "react-icons/ai";
 import { AiOutlinePoweroff } from "react-icons/ai";
 import { db } from "./config/firebase";
 import { auth as firebaseAuth } from './config/firebase';
-import { onAuthStateChanged, getAuth } from 'firebase/auth';
+import { onAuthStateChanged } from 'firebase/auth';
 import Modal from './Modal.jsx';
 import {
   collection,
   query, 
-  orderBy, 
   onSnapshot,
-  getDocs,
   getDoc,
-  addDoc,
   updateDoc,
   deleteDoc,
   doc,
@@ -41,6 +38,7 @@ export default function Game({ auth, gameId, gameCode }) {
 
   const [error, setError] = useState('');
 
+  let newGame = true;
 
   let userEmail = "";
 
@@ -48,64 +46,119 @@ export default function Game({ auth, gameId, gameCode }) {
     const unsubscribe = onAuthStateChanged(firebaseAuth, (currentUser) => {
       setUser(currentUser);
 
+      console.log("User set on Auth State Changed in Game:", currentUser);
+
       // Fetch players for the current game
+
       const fetchPlayers = async () => {
-        console.log(userEmail);
+          if (!gameId) {
+
+              console.error("No game id found on Auth State Changed in Game when fetching players.");
+              return;
+          }
+
+          if (!userEmail) {
+
+            console.error("No user email found in Game when fetching players on Auth State Changed.");
+            return;
+          }
+
           try {
               const response = await axios.get(`/api/game/${gameId}/${userEmail}/players`);
 
-              if (!response.data.players.some(item => item.email === userEmail)) {
-                  alert(`User left the game. Redirecting to home screen...`);
-
-                  router.visit('/');
-
-                  return;
-
+              if (response.data.players === "No players. Game room not found.") {
+                  console.log("No players. Game room not found.");
+                  if(!newGame) {
+                      console.log("User had left the game on Auth State Changed. Redirecting from Game page to home screen...");
+                      router.visit('/');
+                      newGame = false;
+                      return;
+                  }
               }
 
-              console.log(response);
-              setPlayers(response.data.players);
-          } catch (error) {
-              console.error('Error fetching players:', error);
-              router.visit('/');
+              if (!response.data.players.some(item => item.email === userEmail)) {
+                  console.log(`User had left the game on Auth State Changed. Redirecting from Game page to home screen...`);
+                  if(!newGame) {
+                    console.log("User had left the game on Auth State Changed. Redirecting from Game page to home screen...");
+                    newGame = false;
+                    router.visit('/');
+                    return;
+                  }
+              }
 
+              setPlayers(response.data.players);
+
+              console.log("Users fetched successfully in Game page on Auth State Changed:", response.data.players);
+          } catch (error) {
+              console.error('Error fetching players in Game page on Auth State Changed:', error);
+              if(!newGame) {
+                console.log("User had left the game on Auth State Changed. Redirecting from Game page to home screen...");
+                newGame = false;
+                router.visit('/');
+                return;
+              }
           }
       };
 
       // Fetch round for the current game
       const fetchRound = async () => {
+          if (!gameId) {
+
+              console.error("No gameId found in Game when fetching round on Auth State Changed.");
+              return;
+          }
+
+          if (!userEmail) {
+
+            console.error("No user email found in Game when fetching round on Auth State Changed.");
+            return;
+          }
         try {
             const response = await axios.post(`/api/round/${gameId}/${userEmail}`);
 
             if (!response.data.round) {
-              //setRound(0);
             } else {
               setRound(response.data.round);
+              console.log("Round fetched successfully in Game page on Auth State Changed:", response.data.round);
             }
         } catch (error) {
-            console.error('Error fetching players:', error);
+            console.error('Error fetching players on Auth State Changed in Game page:', error);
         }
     };
 
       const q = query(collection(db, 'gameRooms'))
       onSnapshot(q, (querySnapshot) => {
+
+        console.log("Change in firebase detected in Game page on Auth State Changed:", querySnapshot);
         fetchPlayers();
         fetchRound();
+
+        setTimeout(() => {
+          newGame = false;
+        }, 1000);
       })
 
     userEmail = auth?.user?.email || currentUser?.email || "";
 
     if (!userEmail) {
+      console.error("No user email found in Game when fetching players on Auth State Changed. Redirecting to login...");
       router.visit('/login');
     }
 
     const fetchAdmin = async () => {
+
+      if (!userEmail) {
+
+        console.error("No user email found in Game when fetching фвьшт on Auth State Changed.");
+        return;
+      }
       try {
           const response = await axios.get(`/api/game/${userEmail}/admin`);
 
           setAdmin(response.data);
+          console.log("Admin fetched successfully in Game page on Auth State Changed:", response.data);
       } catch (error) {
-          console.error('Error fetching players:', error);
+          console.error('Error fetching players on Auth State Changed:', error);
       }
     };
 
@@ -123,54 +176,90 @@ export default function Game({ auth, gameId, gameCode }) {
 
   useEffect(() => {
 
+    setTimeout(() => {
+      newGame = false;
+    }, 1000);
+
     // Fetch players for the current game
-  const fetchPlayers = async () => {
+    const fetchPlayers = async () => {
+      if (!gameId) {
+
+        console.error("No gameId found in Game when fetching players on page load.");
+        return;
+    }
+
+    if (!userEmail) {
+
+      console.error("No user email found in Game when fetching players on page load.");
+      return;
+    }
       try {
           const response = await axios.get(`/api/game/${gameId}/${user?.email}/players`);
 
           setPlayers(response.data.players);
 
-          console.log(response.data.players);
+          console.log('Players feetched successfully on page load in Game page:', response.data.players);
       } catch (error) {
-          console.error('Error fetching players:', error);
+          console.error('Error fetching players on page load i n Game page:', error);
       }
   };  
 
-    fetchPlayers();
+  fetchPlayers();
 
     // Fetch round for the current game
-  const fetchRound = async () => {
+    const fetchRound = async () => {
+      if (!gameId) {
+
+        console.error("No gameId found in Game when fetching round on page load.");
+        return;
+    }
+
+    if (!userEmail) {
+
+      console.error("No user email found in Game when fetching round on page load.");
+      return;
+    }
+
     try {
         const response = await axios.post(`/api/round/${gameId}/${user?.email}`);
 
         if (!response.data.round) {
-          //setRound(0);
         } else {
           setRound(response.data.round);
+          console.log("Round fetched successfully on page load in Game page:", response.data.round);
         }
     } catch (error) {
-        console.error('Error fetching players:', error);
+        console.error('Error fetching players on page load in game page:', error);
     }
 };
-
 
     fetchRound();
 
 
   }, [round]);
 
-  console.log(user, auth?.user, 'current user');
-
-
   const startNewRound = async () => {
       userEmail = auth?.user?.email || user?.email || "";
+
+      if (!gameId) {
+
+          console.error("No gameId found in Game when starting new round.");
+          return;
+      }
+
+      if (!userEmail) {
+
+        console.error("No user email found in Game when starting new round.");
+        return;
+      }
 
       try {
           const response = await axios.post(`/api/games/${gameId}/${userEmail}/${round + 1}/rounds`);
           setPlayers(response.data.players);
 
-          console.log(response);
           setRound(round + 1);
+
+          console.log('New round started successfully in Game page:', response.data.players);
 
           try {
             // Get the reference to the game document
@@ -181,21 +270,34 @@ export default function Game({ auth, gameId, gameCode }) {
               round: round + 1,
             });
 
-          } catch (err) {
-            console.error('Error setting round for game:', err);
-          }
+            console.log('New round started successfully in Firebase in Game page:', response.data.players);
 
-          setRound(round + 1);
+          } catch (err) {
+            console.error('Error setting round for game in firebase after starting new round :', err);
+          }
       } catch (error) {
-          console.error('Error starting new round:', error);
+          console.error('Error starting new round in Game page:', error);
       }
   };
 
   const leaveGame = async () => {
     userEmail = auth?.user?.email || user?.email || "";
-    console.log(auth);
+
+    if (!gameId) {
+
+        console.error("No gameId found in Game when leaving game.");
+        return;
+    }
+
+    if (!userEmail) {
+
+      console.error("No user email found in Game when leaving game.");
+      return;
+    }
       try {
           await axios.delete(`/api/game/${gameId}/${userEmail}/leave`);
+
+          console.log('Left game successfully in Game page:', gameId);
 
           try {
             // Get the reference to the game document
@@ -217,23 +319,32 @@ export default function Game({ auth, gameId, gameCode }) {
                 await updateDoc(gameDocRef, {
                   players: updatedPlayers,
                 });
-          
-                console.log('Player removed from the game successfully');
+
+                setPlayers([]);
+                setAdmin("");
+                setRound("");
+                setGame([]);
+                setUser(null);
+                setIsAdmin(false);
+                setNewGameCode('');
+                setJoinGameCode('');
+
+                console.log('Player removed from the game in Firebase successfully on leaving game');
           
                 // Check if there are no more players in the game
                 if (updatedPlayers.length === 0) {
                   // Delete the game document
                   await deleteDoc(gameDocRef);
-                  console.log('Game document deleted as there are no more players');
+                  console.log('Game document deleted as there are no more players on leaving game');
                 }
               } else {
-                console.log('Player does not exist in the game');
+                console.log('Player does not exist in the game on leaving game');
               }
             } else {
-              console.error('Game document does not exist');
+              console.error('Game document does not exist on leaving game');
             }
           } catch (err) {
-            console.error('Error removing player from game:', err);
+            console.error('Error removing player from game on leaving game:', err);
           }
           
 
@@ -246,13 +357,25 @@ export default function Game({ auth, gameId, gameCode }) {
 
   const handleJoinGameSubmit = async (e) => {
     e.preventDefault();
+
+    if (!gameCode) {
+
+        console.error("No game code found in Game when joining game from Game page.");
+        return;
+    }
+
+    if (!userEmail) {
+
+      console.error("No user email found in Game when joining game from Game page.");
+      return;
+    }
     try {
       // Send request to backend to join the game
       try {
         const response = await axios.post('/api/join-game', { gameCode: gameCode, userEmail: userEmail });
         
         // Handle successful response
-        console.log('Successfully joined the game:', response.data);
+        console.log('Successfully joined the game from Game page:', response.data);
         
         // Reset the input field after successful submission
         setJoinGameCode('');
@@ -277,15 +400,15 @@ export default function Game({ auth, gameId, gameCode }) {
                 players: updatedPlayers,
               });
         
-              console.log('User added to the game successfully');
+              console.log('User added to the game in Firebase successfully when joining game from Game page');
             } else {
-              console.log('User already exists in the game');
+              console.log('User already exists in tFirebase he game when joining game from Game page');
             }
           } else {
-            console.error('Game document does not exist');
+            console.error('Game document does not exist when joining game from Game page');
           }
         } catch (err) {
-          console.error('Error updating game document:', err);
+          console.error('Error updating game document when joining game from Game page:', err);
         }
 
         // Redirect to the game window
@@ -295,24 +418,20 @@ export default function Game({ auth, gameId, gameCode }) {
         // Handle error response
         if (error.response) {
           // Server responded with a status other than 200 range
-          console.error('Error response:', error.response.data);
+          console.error('Error response when joining game from Game page:', error.response.data);
           setError(error.response.data.error || 'Unknown error');
         } else if (error.request) {
           // Request was made but no response received
-          console.error('Error request:', error.request);
-          alert('No response received from the server. Please try again later.');
+          console.error('Error response when joining game from Game page:', error.request);
         } else {
           // Something else happened in making the request
-          console.error('Error message:', error.message);
-          alert(`Error joining game: ${error.message}`);
+          console.error('Error response when joining game from Game page:', error.message);
         }
       }      
-      
-      console.log('Join game response:', response.data);
-      
+            
     } catch (error) {
       // Handle error if joining game fails
-      console.error('Error joining game:', error.response.data);
+      console.error('Error response when joining game from Game page:', error.response.data);
     }
   };
 
@@ -339,12 +458,6 @@ export default function Game({ auth, gameId, gameCode }) {
               <AiOutlineSend />
             </button>
           }
-          
-          {/* {isUserInPlayers() && (
-            <button onClick={leaveGame} className="bg-red-500 text-white px-4 py-2 rounded-md">
-              <AiOutlinePoweroff />
-            </button>
-          )} */}
 
           {isUserInPlayers() && (
             <button onClick={() => setIsModalOpen(true)} className="bg-red-500 text-white px-4 py-2 rounded-md">
@@ -385,7 +498,14 @@ export default function Game({ auth, gameId, gameCode }) {
           leaveGame();
         }}
         title="Confirm Leave Game"
-        message="Are you sure you want to leave the game? As the admin, leaving will end the game for everyone."
+        message={
+          <span>
+            Are you sure you want to leave the game?
+            {admin.isAdmin && (
+              <span> As you are the admin the game will be deleted.</span>
+            )}
+          </span>
+        }
       />
     </AuthenticatedLayout>
   );
